nitter_requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
aiostream==0.5.2
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.12.2
boto3==1.34.7
botocore==1.34.7
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
distro==1.9.0
fastapi==0.105.0
frozenlist==1.4.1
grpclib==0.4.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httpx==0.26.0
hyperframe==6.0.1
idna==3.6
importlib-metadata==7.0.1
jmespath==1.0.1
lxml==4.9.4
markdown-it-py==3.0.0
mdurl==0.1.2
modal==0.56.4376
multidict==6.0.4
ntscraper==0.3.8
openai==1.6.1
protobuf==4.25.1
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
rich==13.7.0
s3transfer==0.10.0
sigtools==4.0.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
starlette==0.27.0
synchronicity==0.5.3
tblib==3.0.0
toml==0.10.2
tqdm==4.66.1
typer==0.9.0
types-certifi==2021.10.8.3
types-toml==0.10.8.7
typing_extensions==4.9.0
urllib3==2.0.7
watchfiles==0.21.0
yarl==1.9.4
zipp==3.17.0
